move analytics into visual editor provider
